math in place

+++ src/main/java/com/eppo/sdk/helpers/bandit/BanditEvaluator.java
@@ -22,7 +22,7 @@ public class BanditEvaluator {
     ) {
         String modelName = modelParameters != null ? modelParameters.getModelName() : "random";
         BanditModel model = BanditModelFactory.build(modelName);
-        Map<String, Float> actionWeights = model.weighActions(modelParameters, actions, subjectAttributes);
+        Map<String, Double> actionWeights = model.weighActions(modelParameters, actions, subjectAttributes);
         List<String> shuffledActions = shuffleActions(actions.keySet(), experimentKey, subjectKey);
         return generateVariations(shuffledActions, actionWeights, subjectShards);
     }
@@ -40,12 +40,12 @@ public class BanditEvaluator {
         return Shard.getShard(experimentKey+"-"+subjectKey+"-"+actionKey, SHUFFLE_SHARDS);
     }
 
-    private static List<Variation> generateVariations(List<String> shuffledActions, Map<String, Float>  actionWeights, int subjectShards) {
+    private static List<Variation> generateVariations(List<String> shuffledActions, Map<String, Double>  actionWeights, int subjectShards) {
 
         final AtomicInteger lastShard = new AtomicInteger(0);
 
         List<Variation> variations = shuffledActions.stream().map(actionName -> {
-            float weight = actionWeights.get(actionName);
+            double weight = actionWeights.get(actionName);
             int numShards = Double.valueOf(Math.floor(weight * subjectShards)).intValue();
             int shardStart = lastShard.get();
             int shardEnd = shardStart + numShards;
