+++ src/main/java/com/eppo/sdk/helpers/bandit/BanditModel.java
@@ -7,6 +7,6 @@ import java.util.Map;
 
 public interface BanditModel {
 
-    Map<String, Float> weighActions(BanditParameters parameters, Map<String, EppoAttributes> actions, EppoAttributes subjectAttributes);
+    Map<String, Double> weighActions(BanditParameters parameters, Map<String, EppoAttributes> actions, EppoAttributes subjectAttributes);
 
 }
